requirements.txt
Flask==2.1.2
tensorflow==2.8.0
Pillow==8.4.0
gunicorn==20.1.0
